Changed the toggle from video to gameplay
