+++ Assets/Scripts/UI/VideoSettingsBehaviour.cs
@@ -30,11 +30,6 @@ public class VideoSettingsBehaviour : MonoBehaviour
 
     [SerializeField] private Toggle _goreToggleButton;
 
-    [SerializeField] private Toggle _controllerToggleButton;
-
-    private bool _isPreventingUIChange = false;
-    private WaitForEndOfFrame _waitToAllowUIChange = new WaitForEndOfFrame();
-
     /// <summary>
     /// set up references
     /// </summary>
@@ -54,15 +49,6 @@ public class VideoSettingsBehaviour : MonoBehaviour
 
         _subtitleToggleButton.isOn = SaveManager.Instance.GetGameSaveData().IsSubtitlesOn;
         _goreToggleButton.isOn = SaveManager.Instance.GetGameSaveData().IsGoreOn;
-
-        if (_controllerToggleButton.isOn != UiManager.IsUsingController)
-        {
-            _isPreventingUIChange = true;
-
-            StartCoroutine(PreventUISwap());
-
-            _controllerToggleButton.isOn = !_controllerToggleButton.isOn;
-        }
     }
 
     /// <summary>
@@ -94,27 +80,5 @@ public class VideoSettingsBehaviour : MonoBehaviour
     public void ToggleGoreSetting()
     {
         SaveManager.Instance.GetGameSaveData().IsGoreOn = _goreToggleButton.isOn;
-    }
-
-    /// <summary>
-    /// Updates the UI in the game to reflect controller or keyboard inputs
-    /// </summary>
-    public void ToggleControllerSetting()
-    {
-        if (!_isPreventingUIChange)
-        {
-            UiManager.Instance.SwapInput();
-        }
-    }
-
-    /// <summary>
-    /// This is meant to prevent the UI from swapping because Unity's system means that this will be called
-    /// on a value change
-    /// </summary>
-    /// <returns></returns>
-    private IEnumerator PreventUISwap()
-    {
-        yield return _waitToAllowUIChange;
-        _isPreventingUIChange = false;
     }
 }
